requirements.txt
# Core LLM orchestration stack
langchain>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0

# Provider integrations (choose based on enabled providers)
langchain-openai>=0.2.0,<0.3.0
langchain-anthropic>=0.3.0,<0.4.0

# Service runtime
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.8.0,<3.0.0
PyYAML>=6.0.0,<7.0.0
httpx>=0.27.0,<1.0.0

# Optional utilities for reliability/logging
tenacity>=9.0.0,<10.0.0
